Initial Skeleton Page with Simple Nav + Router
Dont know what to use for the shop products. Can either create my own
inventory of items, and locally store the images and stock on a JSON.
Or, I can fetch the data from elsewhere. I initially tried fetching
Riot items data, but if I go this route, It will look decently
complicated. Why? because they store their item descriptions with text
in html format, with their own unique tags. So I will need 2 things:
A way to render the html, and to sanitize the html.

Also their items are stored as an object with keys, each key is its own
item object.

Its definetely doable, but seems kind of complicate... But I might want
to do something with League Items in the future, so I might as well
take a closer look to this approach. Lets see if we can find more info
on how to do this/ alternatives