Simple Layout League Shop with Sort and Filter
Need to implement Cart/Checkout.js and also think about how its state
is handled in App.js

Below, I will copy paste the comments regarding this in App.js:

TODO need to add cartActions to Shop, so that we can further pass
it down to <Item>..
Also we need to think about how we want to handle adding an item to
the cart. Remmember each item has a lot of random keys from when we
fetch the data, and we need a way to keep track of items in cart
AKA, we prob need to add an 'quantity' key or something to
each item obj and then add that object to the cart, then run some
sort of reducer that handles each object add.
We can also instead, just create a new obj that takes only the
important keys from item, then add the 'quantity' key to that new obj

TODO TODO
After thinking for a bit, i think its best that the shop continues to
load items like it is right now. Maybe simplify the item object, so that
it only contains the key/values that we use
Regardless, when you add an item to the cart, its should be a added as
a cartItem instead of an item. cartItem = {...item, quantity: 0}
Also rename Item everywhere to Product, so Product.js and cartProduct
and product. Also, we need 2 different displays depending on if the
product is displayed on the shop page, or if its displayed on the
product specific page. We can do a simple
{isProductDisplay ? '': description: product.description} on Product.js
(currently Item.js)
Or if not, then create a second component like ProductDisplay for when
its displayed on its prodct page.
Shouldn't have too much trouble anyways, since the css for each product
is in Shop.css, and when we make ProductPage.js, it will load css from
ProductPage.css