Added simple error & loading state
TODO TODO
Will now revamp the app, and throw away all the league-related info,
and instead point this to a shop like https://fakestoreapi.com/products.
Will also fix the custom hook so it returns the 3 states mentioned
earlier.
Refer to the 2nd last commit for more info on what is left TODO.