Switched store items to simple use of fakeStoreApi
TODO TODO
Check if theres any errors in the filter/sort logic.
Also fix the cartActions logic. ALL actions require you to recalulate
the total cost of a certain item (including its own quanitty). However,
I have only included this logic in a few of the handlers.
Also, I should be store the TOTAL COST of ALL the items somewhere in
the cart... currently the cart simply holds the items, and never keeps
track of this... It is only on the Cart.js page, where it calculates
the cart's total cost. Yet, it still doesnt store this cost anywhere.

TODO TODO TODO
I have to switch to using the Context API anyways, but I figure that
having a solid fully working program to begin with before switching,
would be better. But maybe this is wrong, and instead of fixing all the
things Ive listed, I can just start switching to the context api, which
should maybe make this a little easier.
Regardless, there are 2 function/logic things to fix from what Ive
listed: The cart and the handlers cost calculating logic, as well as
the filter/sort logic. Think context doesnt make sense there, but
maybe useReduce does ??

ALSO TODO TODO
Switch to using the same fakeStoreApi to retrieve the number of
categories, and their names, in order to create our category filter
bar on the left.
Also, I should maybe switch to using the API for doing the actual
sorting and filtering too... not sure how this would look tho.