A lot of organization & styling changes
-Separated Shop JSX into 3 separate components, for better readability.
-Fixed shop items looking super unaligned, by creating a fixed size
container where the images will be centered on, based on a maximum
height. This makes it so that all the image containers take up the
same space, so the titles and images for all items end up horizontally
aligned with one another.
-Changed the cart item on the navbar, so that each quantity of an item
is counted towards the total items in the cart. So now if you add 3 of
one item, it will say 'Cart 3' instead of 'Cart 1'.
- A LOT of other styling changes for the look & feel of the app, for
example changes for: all the items in the shop, items in the cart,
sidebar, and options area