Added cart loading state, and loading spinner
Use loading state and spinner when Shop loads all items, or when
ItemDetail needs to fetch for a specific item if it hasnt already
been loaded from the shop.

This helps to simulate the network requests that would be nomrally used
in a shopping app, and how that would feel for the user.

Loading state was added to cartContext only when an item is added to
the cart. I still have to think about whether or not I want this in the
handleDelete and handleEditQuantity. This is because for some actions,
some shopping apps prefer to immediately make the change when the user
clicks, and only LATER adjust when the request returns. So I have to
think about this, because I think I prefer for the remove and
editQuanity functions to use this type of approach which is
'action-first, listen to request and then adjust after it comes back'

ALSO:
-changed the Roboto font default import to an import with specifc sizes
-Added an svg icon for removing item from cart

TODO TODO
in css, maybe give everything a margin and padding of 0, and then
adjust later what needs to be adjusted. Becasuse Im running into a lot
of scenarios where buttons, inputs, headings, <a> tags, and other
random html elements have specific margins that I keep having to
specifically set to 0