Mobile support - a LOT added. + Cart Icon
Switched to using Cart icon. In the future, I might change some of the
other navbar links to also be icons. Or maybe just change them for
mobile ?

Some of the big mobile changes added:
-Navbar is now sticky
-Links on the right now show below logo
-Inside shop, options stacks vertically instead of horizontally
-Item detail shows the Image at the top, then the details below
-Cart 'Price Title' text disappears
-Each cart item's total shows below the rest of the items detail. Extra
small 'total' text is displayed

Also changed a lot of small details like paddings, gaps, and font sizes.

TODO TODO
Add effect to cart when an item is added or removed. Add small homepage.
Also some big things to think about for the future:
maybe rethink the hamburger menu inside of shop.js
maybe change the navbars layout overall to be completely different.
For example, If the page had a lot of links, the current navbar wouldnt
work.